pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["Derek <derekzongnanchen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-ollama = "^0.2.0"
langchain-core = "^0.3.0"
bs4 = "^0.0.2"
langgraph = "^0.2.21"
tiktoken = "^0.7.0"
docx2txt = "^0.8"
ipython = "^8.27.0"
opencv-python = "^4.10.0.84"
langfuse = "^2.51.2"
tavily-python = "^0.5.0"
black = "^24.8.0"
pillow = "^10.4.0"
grandalf = "^0.8"
unstructured = {extras = ["docx"], version = "^0.15.13"}
langchain-chroma = "^0.1.4"
pysqlite3-binary = "^0.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
